hotel_project/hotel_con/requirements.txt
asgiref==3.6.0
da-vinci==0.3.0
Django==4.1.5
django-admin-thumbnails==0.2.6
django-jazzmin==2.6.0
django-thumbnails==0.7.0
Pillow==9.4.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
